Move Markdown converter to Markdown namespace
