+++ packages/framework/src/Framework/Services/MarkdownService.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Services;
 
-use Hyde\Framework\Actions\MarkdownConverter;
 use Hyde\Framework\Concerns\Internal\SetsUpMarkdownConverter;
 use Hyde\Helpers\Features;
 use Hyde\Markdown\Contracts\MarkdownPostProcessorContract as PostProcessor;
 use Hyde\Markdown\Contracts\MarkdownPreProcessorContract as PreProcessor;
+use Hyde\Markdown\MarkdownConverter;
 use Hyde\Pages\DocumentationPage;
 use League\CommonMark\Extension\DisallowedRawHtml\DisallowedRawHtmlExtension;
 use League\CommonMark\Extension\HeadingPermalink\HeadingPermalinkExtension;
